Move components to own directory and code improvements
