+++ src/player.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import MainTable from './player-table'
+import PlayerTable from './components/player-table'
 
 class PlayerApp extends React.Component {
 	constructor(props) {
 		super(props);
 		this.answers = {};
-		this.randPositions = {};
+		this.randPositions = new Set(); // set of randomly chosen cells that need to be answered
 		this.submitAnswer = this.submitAnswer.bind(this);
 		this.handleSubmit = this.handleSubmit.bind(this);
 		this.handleNewAnswer = this.handleNewAnswer.bind(this);
@@ -33,7 +33,7 @@ class PlayerApp extends React.Component {
 			element.forEach(question => {
 				// randomly selected questions
 				if (this.props.randCount !== 0) {
-					if (this.randPositions.hasOwnProperty(counter)) {
+					if (this.randPositions.has(counter)) {
 						this.submitAnswer(question.id, counter);
 					}
 				}
@@ -64,8 +64,8 @@ class PlayerApp extends React.Component {
 		while (selectCount < this.props.randCount) {
 			const position = Math.floor(Math.random() * totalCells);
 
-			if (!this.randPositions.hasOwnProperty(position)) {
-				this.randPositions[position] = true;
+			if (!this.randPositions.has(position)) {
+				this.randPositions.add(position);
 				selectCount++;
 			}
 		}
@@ -73,7 +73,7 @@ class PlayerApp extends React.Component {
 
 	render() {
 		// randomize which entries are blank if the creator says more than 0 should be random and this is the first render of the table
-		if (Object.entries(this.randPositions).length === 0 && this.randPositions.constructor === Object && this.props.randCount !== 0) {
+		if (this.randPositions.size === 0 && this.props.randCount !== 0) {
 			this.randomize();
 		}
 
@@ -84,7 +84,7 @@ class PlayerApp extends React.Component {
 				<form onSubmit={this.handleSubmit}>
 					<table>
 						<tbody>
-							<MainTable
+							<PlayerTable
 								dimensions={this.props.dimensions}
 								qset={this.props.qset}
 								parentAnswers={this.answers}
